Make member_discount dynamic.
We want to automatically get the correct member discount from the
membership status, even if that changes after a registration. For this
we replace member_discount by the property effective_member_discount
that always calculates the discount ad-hoc.

We keep the column member_discount for now for documentation purposes,
but it won't be shown.