feat: Add transfer date to person payments.
This is nice to have for auditing purposes. The migration extracts dates
from the current work arounds for not having this:
 * setting start and end to the same value for non-membership-payments
 * Adding a comment containing the date.
This is a bit of a heuristic, but should work well.

Since we do not need to workaround anymore, we make start and end
optional for non-membership-payments.

The transfer date is now mandatory, however, we leave it optional for
existing payments (since not all of those are known if they are too far
in the past).