feat: Also track expenses via registration payments.
An expense that needs to be reimbursed is nothing different than a
negative payment. It is also convenient to track them in the db, since
they are needed to calculate the correct debit.

If we start tracking the expenses, we might as well go all the way and
track the category with the aim of generating the bank report of the
event from the db.